requirements.txt
piexif
Pillow
pytest